commited the code of backend (including tests)

+++ backend/src/main/resources/db/migration/V2__create_categories_table.sql
@@ -2,7 +2,7 @@ CREATE TABLE categories (
                             id BIGINT AUTO_INCREMENT PRIMARY KEY,
                             name VARCHAR(64) NOT NULL,
                             parent_id BIGINT,
-                            level TINYINT NOT NULL DEFAULT 1,
+                            level INT NOT NULL DEFAULT 1,
                             sort_order INT NOT NULL DEFAULT 0,
                             icon_url VARCHAR(512),
                             status VARCHAR(16) NOT NULL DEFAULT 'ACTIVE',
